seperated page from component
